Separated the app into two flavors
