+++ app/src/admin/res/layout/activity_login.xml
@@ -23,9 +23,11 @@
   ~ SOFTWARE.
   -->
 
-<manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.twoplaytech.drbetting">
+<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    tools:context=".admin.LoginActivity">
 
-    <application></application>
-
-</manifest>
+</androidx.constraintlayout.widget.ConstraintLayout>
