+++ app/src/admin/res/mipmap-anydpi-v26/ic_launcher_round.xml
@@ -23,9 +23,7 @@
   ~ SOFTWARE.
   -->
 
-<manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.twoplaytech.drbetting">
-
-    <application></application>
-
-</manifest>
+<adaptive-icon xmlns:android="http://schemas.android.com/apk/res/android">
+    <background android:drawable="@color/ic_launcher1_background"/>
+    <foreground android:drawable="@mipmap/ic_launcher_foreground"/>
+</adaptive-icon>
