+++ app/src/admin/res/values/strings.xml
@@ -1,4 +1,3 @@
-<?xml version="1.0" encoding="utf-8"?>
 <!--
   ~ MIT License
   ~
@@ -23,9 +22,4 @@
   ~ SOFTWARE.
   -->
 
-<manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.twoplaytech.drbetting">
-
-    <application></application>
-
-</manifest>
+<resources></resources>
