+++ app/src/client/AndroidManifest.xml
@@ -1,0 +1,59 @@
+<!--
+  ~ MIT License
+  ~
+  ~ Copyright (c)  2021 Damjan Miloshevski
+  ~
+  ~ Permission is hereby granted, free of charge, to any person obtaining a copy
+  ~ of this software and associated documentation files (the "Software"), to deal
+  ~ in the Software without restriction, including without limitation the rights
+  ~ to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+  ~ copies of the Software, and to permit persons to whom the Software is
+  ~ furnished to do so, subject to the following conditions:
+  ~
+  ~ The above copyright notice and this permission notice shall be included in all
+  ~ copies or substantial portions of the Software.
+  ~
+  ~ THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+  ~ IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+  ~ FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+  ~ AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+  ~ LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+  ~ OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+  ~ SOFTWARE.
+  -->
+
+<manifest xmlns:android="http://schemas.android.com/apk/res/android"
+    package="com.twoplaytech.drbetting">
+
+    <application
+        android:name="com.twoplaytech.drbetting.App"
+        android:allowBackup="true"
+        android:icon="@mipmap/ic_launcher"
+        android:roundIcon="@mipmap/ic_launcher_round"
+        android:supportsRtl="true"
+        android:label="@string/app_name"
+        android:theme="@style/Theme.DrBetting">
+        >
+        <activity
+            android:name="com.twoplaytech.drbetting.SplashActivity"
+            android:noHistory="true"
+            android:theme="@style/SplashScreenTheme">
+            <intent-filter>
+                <action android:name="android.intent.action.MAIN" />
+
+                <category android:name="android.intent.category.LAUNCHER" />
+            </intent-filter>
+        </activity>
+        <activity
+            android:name="com.twoplaytech.drbetting.AppInfoActivity"
+            android:theme="@style/Theme.DrBetting.NoActionBar" />
+        <activity
+            android:name="com.twoplaytech.drbetting.ui.settings.SettingsActivity"
+            android:label="@string/title_activity_settings"
+            android:theme="@style/Theme.DrBetting.NoActionBar" />
+        <activity
+            android:name="com.twoplaytech.drbetting.MainActivity"
+            android:screenOrientation="portrait"
+            android:label="@string/app_name"/>
+    </application>
+</manifest>
