+++ app/src/client/res/mipmap-anydpi-v26/ic_launcher.xml
