+++ app/src/client/res/mipmap-hdpi/ic_launcher.png
