+++ app/src/client/res/mipmap-hdpi/ic_launcher_foreground.png
