+++ app/src/client/res/mipmap-mdpi/ic_launcher.png
