+++ app/src/client/res/mipmap-xhdpi/ic_launcher.png
