+++ app/src/client/res/mipmap-xhdpi/ic_launcher_foreground.png
