+++ app/src/client/res/mipmap-xxhdpi/ic_launcher.png
