+++ app/src/client/res/mipmap-xxhdpi/ic_launcher_foreground.png
