+++ app/src/client/res/mipmap-xxxhdpi/ic_launcher.png
